changed tipe of AnwserTimeOut

+++ Trivia/Trivia/Response.h
@@ -62,5 +62,5 @@ typedef struct GetRoomStateResponse {
 	bool hasGameBegun;
 	std::vector<std::string> players;
 	unsigned int questionCount;
-	float answerTimeOut;
+	unsigned int answerTimeOut;
 } GetRoomStateResponse;
